break out client components for usable page titles
